Deno env, deno fmt, deno compile, etc

+++ src/init.ts
@@ -1,22 +1,13 @@
 import type { Config } from "./config.ts";
 import { NODES_PATH } from "./nodes.ts";
 import { BLOCKSDIR } from "./block.ts";
-import { join } from "https://deno.land/std/path/mod.ts";
-import { existsSync } from "https://deno.land/std/fs/mod.ts";
+import { existsSync } from "https://deno.land/std@0.224.0/fs/mod.ts";
+import fs from "node:fs";
 
 function init(config: Config): void {
-  if (!existsSync(join(new URL('.', import.meta.url).pathname, "../files"))) {
-    Deno.mkdir(join(new URL('.', import.meta.url).pathname, "../files"), { recursive: true });
-  }
-  if (!existsSync(NODES_PATH)) {
-    Deno.writeFileSync(
-      NODES_PATH,
-      new TextEncoder().encode(JSON.stringify(config.bootstrapNodes)),
-    );
-  }
-  Deno.mkdir(BLOCKSDIR).catch((err) => {
-    if (!(err instanceof Deno.errors.AlreadyExists)) throw err;
-  });
+	if (!existsSync("files/")) Deno.mkdir("files", { recursive: true });
+	if (!existsSync(NODES_PATH)) Deno.writeFileSync(NODES_PATH, new TextEncoder().encode(JSON.stringify(config.bootstrapNodes)));
+	if (!fs.existsSync(BLOCKSDIR)) Deno.mkdir(BLOCKSDIR);
 }
 
 export default init;
